Refactor shared folder
