+++ podium-gateway/src/main/webapp/app/app.module.ts
@@ -2,20 +2,6 @@ import './vendor.ts';
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { Ng2Webstorage } from 'ng2-webstorage';
-import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
-import {
-    LayoutRoutingModule,
-    PdmMainComponent,
-    NavbarComponent,
-    FooterComponent,
-    ProfileService,
-    PageRibbonComponent,
-    ActiveMenuDirective,
-    ErrorComponent,
-    CompletedComponent
-} from './layouts';
-import { HomeComponent } from './home';
-import { DashboardComponent } from './dashboard';
 import { customHttpProvider } from './blocks/interceptor/http.provider';
 import { PaginationConfig } from './blocks/config/uib-pagination.config';
 import { TypeaheadModule } from 'ng2-bootstrap';
@@ -23,21 +9,29 @@ import { PodiumGatewayRequestModule } from './request/request.module';
 import { PodiumGatewayBbmriBackofficeModule } from './backoffice/bbmri/bbmri-backoffice.module';
 import { PodiumGatewayOrganisationBackofficeModule } from './backoffice/organisation/organisation-backoffice.module';
 import { PodiumGatewayPodiumBackofficeModule } from './backoffice/podium/podium-backoffice.module';
-import { RoleService } from './shared/role/role.service';
 import { PodiumGatewaySharedModule, UserRouteAccessService } from './shared';
 import { PodiumGatewayAdminModule } from './admin/admin.module';
 import { PodiumGatewayAccountModule } from './account/account.module';
 import { BreadcrumbsModule } from './shared/breadcrumbs/breadcrumbs.module';
 import { OrganisationService } from './shared/organisation/organisation.service';
+import { CommonModule } from '@angular/common';
+import { RoleService } from './shared/role/role.service';
+import { PdmMainComponent } from './main/main.component';
+import { PageRibbonComponent } from './shared/profiles/page-ribbon.component';
+import { AppRoutingModule } from './app.route';
+import { HomeComponent } from './home/home.component';
+import { FooterComponent } from './shared/footer/footer.component';
+import { DashboardComponent } from './dashboard/dashboard.component';
 
 @NgModule({
     imports: [
+        CommonModule,
         TypeaheadModule.forRoot(),
         BreadcrumbsModule.forRoot(),
-        BrowserModule,
-        LayoutRoutingModule,
         Ng2Webstorage.forRoot({ prefix: 'pdm', separator: '-'}),
-        PodiumGatewaySharedModule.forRoot(),
+        BrowserModule,
+        AppRoutingModule,
+        PodiumGatewaySharedModule,
         PodiumGatewayAdminModule,
         PodiumGatewayRequestModule,
         PodiumGatewayAccountModule,
@@ -46,18 +40,13 @@ import { OrganisationService } from './shared/organisation/organisation.service'
         PodiumGatewayPodiumBackofficeModule
     ],
     declarations: [
-        PdmMainComponent,
-        HomeComponent,
         DashboardComponent,
-        NavbarComponent,
-        ErrorComponent,
-        CompletedComponent,
+        PdmMainComponent,
         PageRibbonComponent,
-        ActiveMenuDirective,
+        HomeComponent,
         FooterComponent
     ],
     providers: [
-        ProfileService,
         { provide: Window, useValue: window },
         { provide: Document, useValue: document },
         customHttpProvider(),
