+++ podium-gateway/src/main/webapp/app/app.route.ts
@@ -1,17 +1,35 @@
 /*
  * Copyright (c) 2017. The Hyve and respective contributors
  *
- * Licensed under the Apache License, Version 2.0 (the "License");
+ * Licensed under the Apache License, Version 2.0 (the 'License');
  * you may not use this file except in compliance with the License.
  *
  * See the file LICENSE in the root of this repository.
  *
  */
-import { Route } from '@angular/router';
-import { NavbarComponent } from './layouts';
+import { NgModule } from '@angular/core';
+import { RouterModule } from '@angular/router';
+import { homeRoute } from './home/home.route';
+import { dashboardRoute } from './dashboard/dashboard.route';
+import { navbarRoute } from './shared/navbar/navbar.route';
+import { errorRoute } from './shared/error/error.route';
+import { completedRoute } from './shared/completed/completed.route';
 
-export const navbarRoute: Route = {
-    path: '',
-    component: NavbarComponent,
-    outlet: 'navbar'
-  };
+let LAYOUT_ROUTES = [
+    homeRoute,
+    dashboardRoute,
+    navbarRoute,
+    ...errorRoute,
+    ...completedRoute
+];
+
+@NgModule({
+    imports: [
+        RouterModule.forRoot(LAYOUT_ROUTES, { useHash: true })
+    ],
+    exports: [
+        RouterModule
+    ]
+})
+export class AppRoutingModule {
+}
