--- podium-gateway/src/main/webapp/app/layouts/index.ts
@@ -1,22 +1,0 @@
-/*
- * Copyright (c) 2017. The Hyve and respective contributors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- *
- * See the file LICENSE in the root of this repository.
- *
- */
-
-export * from './completed/completed.component';
-export * from './completed/completed.route';
-export * from './error/error.component';
-export * from './error/error.route';
-export * from './main/main.component';
-export * from './footer/footer.component';
-export * from './navbar/navbar.component';
-export * from './navbar/active-menu.directive';
-export * from './profiles/page-ribbon.component';
-export * from './profiles/profile.service';
-export * from './profiles/profile-info.model';
-export * from './layout-routing.module';
