+++ podium-gateway/src/main/webapp/app/main/main.component.ts
@@ -9,8 +9,8 @@
  */
 import { Component, OnInit } from '@angular/core';
 import { Router, ActivatedRouteSnapshot, NavigationEnd } from '@angular/router';
-import { JhiLanguageHelper } from '../../shared';
-import { Principal } from '../../shared/auth/principal.service';
+import { JhiLanguageHelper } from '../shared';
+import { Principal } from '../shared/auth/principal.service';
 
 @Component({
     selector: 'pdm-main',
