+++ podium-gateway/src/main/webapp/app/shared/completed/completed.component.html
@@ -12,8 +12,9 @@
         <div class="message-container">
             <h1>{{completedTitle}}</h1>
             <p [innerHTML]="completedContent"></p>
-            <a *ngIf="isCompletionType(completionTypes.Registration)" href="/#/">Login</a>
-            <a *ngIf="isCompletionType(completionTypes.Request_Submit)" href="/#/requests/my-requests">Go to request overview</a>
+            <a *ngIf="isCompletionType(completionTypes.Registration)" href="/">Login</a>
+            <a *ngIf="isCompletionType(completionTypes.Request_Submit)" href="requests/my-requests">Go to request
+                overview</a>
         </div>
     </div>
     <div class="image-header">
