+++ podium-gateway/src/main/webapp/app/shared/completed/completed.scss
@@ -1,0 +1,32 @@
+/*!
+ * Copyright (c) 2017. The Hyve and respective contributors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ *
+ * See the file LICENSE in the root of this repository.
+ *
+ */
+
+.image-header {
+    width: 300px;
+    height: 200px;
+    margin: 0 auto;
+
+    .logo {
+        height: 400px;
+        width: 400px;
+        display: inline-block;
+        vertical-align: middle;
+        background: url("../../../content/images/bbmri-logo.png") no-repeat center center;
+        background-size: contain;
+    }
+}
+
+.message-container {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    margin-top: 50px;
+}
+
