+++ podium-gateway/src/main/webapp/app/shared/completed/completion-type.ts
