+++ podium-gateway/src/main/webapp/app/shared/error/error.component.ts
