+++ podium-gateway/src/main/webapp/app/shared/error/error.route.ts
@@ -1,0 +1,33 @@
+/*
+ * Copyright (c) 2017. The Hyve and respective contributors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ *
+ * See the file LICENSE in the root of this repository.
+ *
+ */
+import { Routes } from '@angular/router';
+import { UserRouteAccessService } from '../';
+import { ErrorComponent } from './error.component';
+
+export const errorRoute: Routes = [
+    {
+        path: 'error',
+        component: ErrorComponent,
+        data: {
+            authorities: [],
+            pageTitle: 'error.title'
+        },
+        canActivate: [UserRouteAccessService]
+    },
+    {
+        path: 'accessdenied',
+        component: ErrorComponent,
+        data: {
+            authorities: [],
+            pageTitle: 'error.title'
+        },
+        canActivate: [UserRouteAccessService]
+    }
+];
