+++ podium-gateway/src/main/webapp/app/shared/footer/footer.component.ts
