+++ podium-gateway/src/main/webapp/app/shared/index.ts
@@ -34,6 +34,7 @@ export * from './auth/auth-jwt.service';
 export * from './auth/auth.service';
 export * from './auth/principal.service';
 export * from './auth/has-any-authority.directive';
+export * from './auth/user-route-access-service';
 export * from './language/language.constants';
 export * from './language/language.helper';
 export * from './language/language.pipe';
@@ -49,7 +50,6 @@ export * from './organisation-selector/organisation-selector.component';
 export * from './shared-libs.module';
 export * from './shared-common.module';
 export * from './shared.module';
-export * from './auth/user-route-access-service';
 export * from './breadcrumbs/breadcrumbs.module';
 export * from './delivery/delivery';
 export * from './delivery/delivery.service';
@@ -58,3 +58,18 @@ export * from './delivery/delivery-state-options.constants';
 export * from './delivery/delivery-status.constants';
 export * from './overview/overview.service';
 export * from './overview/overview';
+export * from './completed/completed.component';
+export * from './completed/completed.route';
+export * from './error/error.component';
+export * from './error/error.route';
+export * from './not-found/not-found.component';
+export * from './not-found/not-found.route';
+export * from './footer/footer.component';
+export * from './navbar/navbar.component';
+export * from './navbar/navbar.route';
+export * from './navbar/active-menu.directive';
+export * from './authority/authority.constants';
+export * from './authority/authority';
+export * from './profiles/page-ribbon.component';
+export * from './profiles/profile.service';
+export * from './profiles/profile-info.model';
