+++ podium-gateway/src/main/webapp/app/shared/navbar/active-menu.directive.ts
