+++ podium-gateway/src/main/webapp/app/shared/navbar/navbar.route.ts
@@ -1,0 +1,18 @@
+/*
+ * Copyright (c) 2017. The Hyve and respective contributors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ *
+ * See the file LICENSE in the root of this repository.
+ *
+ */
+
+import { Route } from '@angular/router';
+import { NavbarComponent } from './navbar.component';
+
+export const navbarRoute: Route = {
+    path: '',
+    component: NavbarComponent,
+    outlet: 'navbar'
+};
