+++ podium-gateway/src/main/webapp/app/shared/navbar/navbar.scss
@@ -1,0 +1,99 @@
+/*!
+ * Copyright (c) 2017. The Hyve and respective contributors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ *
+ * See the file LICENSE in the root of this repository.
+ *
+ */
+
+
+/* ==========================================================================
+Navbar
+========================================================================== */
+
+$link-color: #ccc;
+
+.navbar-version {
+    font-size: 10px;
+    color: $link-color;
+}
+
+.jh-navbar {
+    background-color: #353d47;
+    padding: .8em 1em;
+    border-radius: 0;
+
+    .dropdown-item {
+        &.active,
+        &.active:focus,
+        &.active:hover {
+            background-color: #353d47;
+        }
+    }
+
+    .dropdown-toggle::after {
+        margin-left: 0.15em;
+    }
+    ul.navbar-nav {
+        padding: 0.5em;
+        .nav-item+.nav-item {
+            margin-left: 1.5rem;
+        }
+
+        .nav-item {
+            color: $link-color;
+        }
+
+        .nav-item .material-icons {
+            margin-right: 7px;
+        }
+    }
+
+    a.nav-link {
+        font-weight: 400;
+    }
+
+    .jh-navbar-toggler {
+        color: #ccc;
+        font-size: 1.5em;
+        padding: 10px;
+        &:hover {
+            color: #fff;
+        }
+    }
+
+    .account-name {
+        margin-right: 5px;
+    }
+}
+
+@media screen and (max-width: 992px) {
+    .jh-logo-container {
+        width: 100%;
+    }
+}
+
+.navbar-title {
+    display: inline-block;
+    vertical-align: middle;
+    color: $link-color;
+}
+
+/* ==========================================================================
+Logo styles
+========================================================================== */
+.navbar-brand {
+    &.logo {
+        padding: 5px 15px;
+        .logo-img {
+            height: 45px;
+            width: 70px;
+            display: inline-block;
+            vertical-align: middle;
+            background: url("../../../content/images/bbmri-logo-image.png") no-repeat center center;
+            background-size: contain;
+        }
+    }
+}
