+++ podium-gateway/src/main/webapp/app/shared/not-found/not-found.component.html
@@ -1,0 +1,4 @@
+<div class="content-wrapper">
+    <h1>Page Not Found</h1>
+    <p>Sorry, but the page you were trying to view does not exist.</p>
+</div>
