+++ podium-gateway/src/main/webapp/app/shared/profiles/page-ribbon.component.ts
