+++ podium-gateway/src/main/webapp/app/shared/shared.module.ts
@@ -13,17 +13,10 @@ import { SessionStorageService } from 'ng2-webstorage';
 import {
     PodiumGatewaySharedLibsModule,
     PodiumGatewaySharedCommonModule,
-    CSRFService,
-    AuthService,
-    AuthServerProvider,
-    AccountService,
     UserService,
-    StateStorageService,
     LoginService,
     LoginModalService,
     MessageService,
-    Principal,
-    HasAnyAuthorityDirective,
     EmailValidatorDirective,
     PasswordValidatorDirective,
     PasswordMatchesDirective,
@@ -38,11 +31,22 @@ import { RequestAccessService } from './request/request-access.service';
 import { RequestReviewPanelComponent } from './request/request-review-panel/request-review-panel.component';
 import { PodiumEventMessageComponent } from './event/podium-event-message.component';
 import { LinkedRequestNotificationComponent } from './linked-request-notification/linked-request-notification.component';
+import { PodiumAuthModule } from './auth/auth.module';
+import { FindLanguageFromKeyPipe } from './language/language.pipe';
+import { ActiveMenuDirective } from './navbar/active-menu.directive';
+import { CompletedComponent } from './completed/completed.component';
+import { ErrorComponent } from './error/error.component';
+import { NavbarComponent } from './navbar/navbar.component';
+import { NotFoundComponent } from './not-found/not-found.component';
+import { ProfileService } from './profiles/profile.service';
+import { RouterModule } from '@angular/router';
 
 @NgModule({
     imports: [
         PodiumGatewaySharedLibsModule,
-        PodiumGatewaySharedCommonModule
+        PodiumGatewaySharedCommonModule,
+        PodiumAuthModule,
+        RouterModule
     ],
     declarations: [
         PodiumLoginComponent,
@@ -55,10 +59,25 @@ import { LinkedRequestNotificationComponent } from './linked-request-notificatio
         PasswordValidatorDirective,
         PasswordMatchesDirective,
         WordLengthValidatorDirective,
-        HasAnyAuthorityDirective,
-        EnumKeysPipe
+        EnumKeysPipe,
+        FindLanguageFromKeyPipe,
+        ActiveMenuDirective,
+        CompletedComponent,
+        ErrorComponent,
+        NavbarComponent,
+        NotFoundComponent,
     ],
-    providers: [],
+    providers: [
+        ProfileService,
+        LoginService,
+        LoginModalService,
+        MessageService,
+        SessionStorageService,
+        UserService,
+        RequestService,
+        RequestAccessService,
+        DatePipe
+    ],
     entryComponents: [PodiumLoginComponent],
     exports: [
         PodiumGatewaySharedCommonModule,
@@ -73,33 +92,14 @@ import { LinkedRequestNotificationComponent } from './linked-request-notificatio
         PasswordValidatorDirective,
         PasswordMatchesDirective,
         WordLengthValidatorDirective,
-        HasAnyAuthorityDirective,
         DatePipe,
-        EnumKeysPipe
+        EnumKeysPipe,
+        ActiveMenuDirective,
+        FindLanguageFromKeyPipe,
     ],
     schemas: [CUSTOM_ELEMENTS_SCHEMA],
 
 })
 export class PodiumGatewaySharedModule {
-    static forRoot(): ModuleWithProviders {
-        return {
-            ngModule: PodiumGatewaySharedModule,
-            providers: [
-                LoginService,
-                LoginModalService,
-                MessageService,
-                AccountService,
-                SessionStorageService,
-                StateStorageService,
-                Principal,
-                CSRFService,
-                AuthServerProvider,
-                AuthService,
-                UserService,
-                RequestService,
-                RequestAccessService,
-                DatePipe
-            ]
-        };
-    }
+
 }
